backed/requirements.txt
asgiref==3.11.0
click==8.3.1
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.40.0
